Solver: Created a new method that will return the possible solutions ordered by their probability of being one. This method is called is 'find_most_likely_solutions'. The way it works is, we first keep track of every possible card the solution can be. Then we run N iterations where we generate a random solution, assign all other cards randomly to players and we check if assigning those cards results in a valid state for the players. If it is then we count that iteration for every card in that solution.
The larger N gets, the more precise the results will be but obviously
the slower the method will be. The goal is to try to optimize the solver
so that we can use larger values for N.
Right now, I set it to 200_000 iterations which runs okay in Release
mode.